utils: Moved cleaning command strings to general
This is work towards #268

The cleaning of command strings can be used within the Dockerfile
context or within the Docker image's 'created_by' key.

- Moved clean_command function and regex string to general.py
- Import clean_command into dockerfile.py
