+++ tern/utils/general.py
@@ -13,6 +13,10 @@ from tern import Version
 from tern.utils import constants
 
 
+# regex strings
+cleaning = '[\t\\\\]'
+
+
 # from https://stackoverflow.com/questions/6194499/pushd-through-os-system
 @contextmanager
 def pushd(path):
@@ -27,6 +31,13 @@ def initialize_names():
     constants.image = constants.image + "_" + str(randint)
     constants.tag = constants.tag + "_" + str(randint)
     constants.container = constants.container + "_" + str(randint)
+
+
+def clean_command(command):
+    '''Given a command string, clean out all whitespaces, tabs and line
+    indentations
+    Leave && alone'''
+    return re.sub(cleaning, '', command).strip()
 
 
 def parse_command(command):
